Bug fixes and improvements

+++ microservizi/Inventario/Inventario.Business/Abstractions/IBusiness.cs
@@ -6,7 +6,7 @@ public interface IBusiness
 {
     // Articoli
     Task CreateArticoloAsync(string nome, string desc, decimal prezzo, int quantita, string SKU, string categoria, int fk_fornitore, CancellationToken cancellationToken = default);
-    Task<ArticoloDto?> ModificaPrezzoArticoloAsync(int id, int nuovoPrezzo, CancellationToken cancellationToken = default);
+    Task<ArticoloDto?> ModificaPrezzoArticoloAsync(int id, decimal nuovoPrezzo, CancellationToken cancellationToken = default);
     Task<ArticoloDto?> ScaricaQuantitaAsync(int id, int quantita, CancellationToken cancellationToken = default);
     Task<ArticoloDto?> ReadArticoloAsync(int id, CancellationToken cancellationToken = default);
     Task<ArticoloDto?> GetSkuAsync(string CodiceSKU, CancellationToken cancellationToken = default);
